♻️ Refactor functions
